refactor(app): optimize internals
