+++ internal/api/general.go
